requirements.txt
python-telegram-bot==20.7
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
pillow==10.1.0
aiohttp==3.9.1
cryptography==41.0.8